requirements.txt
# Core dependencies
numpy>=1.21.0
pandas>=1.3.0
scipy>=1.7.0
scikit-learn>=1.0.0

# Deep learning
torch>=2.0.0
transformers>=4.30.0
sentence-transformers>=2.2.0

# LLM providers
google-generativeai>=0.3.0  # For Gemini
openai>=1.0.0  # For OpenAI
anthropic>=0.7.0  # For Claude

# Configuration
pyyaml>=6.0

# Progress bars and UI
tqdm>=4.65.0

# Experiment tracking (optional)
mlflow>=2.3.0
wandb>=0.15.0

# Evaluation
nltk>=3.8.0

# Optional: for advanced clustering
hdbscan>=0.8.0

# Development
pytest>=7.3.0
black>=23.3.0
flake8>=6.0.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
